requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
neo4j==5.14.1
sqlalchemy==2.0.23

# GraphQL
strawberry-graphql[fastapi]==0.217.0

# Data processing
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
networkx==3.2.1
joblib==1.3.2

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
aioresponses==0.7.6

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
